+++ src/main/java/attendance/domain/Attendance.java
@@ -1,4 +1,4 @@
-package attendance;
+package attendance.domain;
 
 import java.time.DayOfWeek;
 import java.time.LocalDate;
