+++ src/main/java/attendance/domain/HourMinute.java
@@ -1,4 +1,4 @@
-package attendance;
+package attendance.domain;
 
 public record HourMinute(int hour, int minute) {
 }
